// src/models/IEmployeeTranslation.ts
export interface EmployeeTranslations {
  addEmployee: string;
  searchEmployee: string;
  clearEmployee: string;
  editEmployee: string;
  nameEmployee: string;
  phoneEmployee: string;
  statusEmployee: string;
  positionEmployee: string;
  actionEmployee: string;
  noEmployee: string;
  dobEmployee: string;
  genderEmployee: string;
  status1: string;
  status2: string;
  status3: string;
  allAccount: string;
  hasAccount: string;
  noAccount: string;
  selectposition: string;
  selectdob: string;
  male: string;
  female: string;
  saveEmployee: string;
  cancelEmployee: string;
  confirm1: string;
  confirm2: string;
  confirm3: string;
  confirm4: string;
  confirm5: string;
  confirm6: string;
  confirm7: string;
  confirm8: string;
  confirm9: string;
  confirm10: string;
  statusAdd: string;
  statusEdit: string;
  detail: string;
  page: string;
  load: string;
  empty: string;
  export: string;
  noPosition: string;
  delete: string;
  confirm11: string;
  confirm12: string;
  yes: string;
  no: string;
  cv: string;
  skill: string;
  selectSkill: string;
  confirm13: string;
  confirm14: string;
  updateStatus: string;
  only: string;
  filter: string;
  idCard: string;
  project: string;
  editProfile: string;
  description: string;
  infor: string;
  noFile: string;
  failCanvas: string;
  failConvert: string;
  uploadFail: string;
  image: string;
  picture: string;
  pic1: string;
  pic2: string;
}